django-backend/users/serializers.py
from rest_framework import serializers

from users.models import CustomUser as User


class UserSerializer(serializers.ModelSerializer):
    videos = serializers.PrimaryKeyRelatedField(many=True, read_only=True)

    class Meta:
        model = User
        fields = ['id', 'username', 'avatar', 'videos']
