django-backend/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-rest-framework
djangorestframework
pytz
sqlparse
django-redis
redis
schedule
django_cron
gunicorn
dj-rest-auth[with_social]
django-allauth
Pillow